Removed cli-server and moved all code back into src/server.rs
